hive/gui/backend/requirements.txt
Django>=5.2
asgiref==3.8.1
sqlparse==0.5.0
tzdata==2025.1
djangorestframework
django-cors-headers
django-extensions
mysqlclient
